Send OK headers with directory listing and CGI
There seems to be issues that call for a refactoring.  In order to
return the Content-Length, it seems appropriate to first copy the
directory listing and CGI output to a buffer and then return the
length of this buffer.  Further, any error that occurs during the
directory listing or CGI should cause the server to send an internal
error (500) response.